refactor(domain): refactor domain module and aggregate conventional commits validation

+++ driven/repository-sql/src/main/java/com/price/manager/driven/repositories/models/PriceEntity.java
@@ -1,14 +1,23 @@
 package com.price.manager.driven.repositories.models;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
+/**
+ * Entidad JPA para persistencia de precios.
+ */
 @Entity
 @Data
 @Builder
@@ -17,30 +26,30 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class PriceEntity implements Serializable {
 
-	@Id
-	@Column(name = "PRICE_LIST", nullable = false)
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long priceList;
+    @Id
+    @Column(name = "PRICE_LIST", nullable = false)
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long priceList;
 
-	@Column(name = "START_DATE", nullable = false)
-	private LocalDateTime startDate;
+    @Column(name = "START_DATE", nullable = false)
+    private LocalDateTime startDate;
 
-	@Column(name = "END_DATE", nullable = false)
-	private LocalDateTime endDate;
+    @Column(name = "END_DATE", nullable = false)
+    private LocalDateTime endDate;
 
-	@Column(name = "BRAND_ID", nullable = false)
-	private Long brandId;
+    @Column(name = "BRAND_ID", nullable = false)
+    private Long brandId;
 
-	@Column(name = "PRODUCT_ID", nullable = false)
-	private Long productId;
+    @Column(name = "PRODUCT_ID", nullable = false)
+    private Long productId;
 
-	@Column(name = "PRIORITY", nullable = false)
-	private Integer priority;
+    @Column(name = "PRIORITY", nullable = false)
+    private Integer priority;
 
-	@Column(name = "PRICE", nullable = false)
-	private Double price;
+    @Column(name = "PRICE", nullable = false)
+    private BigDecimal price;
 
-	@Column(name = "CURR", nullable = false)
-	private String curr;
+    @Column(name = "CURR", nullable = false)
+    private String curr;
 
 }
